Proj1/requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
langdetect>=1.0.9
